Move completions_from_composite_sources() into the builder
